requirements.txt
# Polymarket Trading Bot Requirements

# Core dependencies
requests>=2.31.0

# Polymarket official client (required for trading)
py-clob-client>=0.17.0

# Web3 dependencies (required by py-clob-client)
web3>=6.0.0
eth-account>=0.9.0

# Utilities
python-dotenv>=1.0.0  # For environment variables
python-dateutil>=2.8.2  # For date parsing
pytz>=2023.3  # For timezone handling

